Update JBig2Image.cs
I encountered an image file where the top row had the predictor flag set. This caused an IndexOutOfRangeException in FastBitSet because it was trying to read row -1. I discovered that this indicates that the bottom row should be used as the predictor. Making this change allowed me to decode the image successfully.